+++ src/MQTTAsyncUtils.c
@@ -2034,7 +2034,7 @@ static int MQTTAsync_completeConnection(MQTTAsyncs* m, Connack* connack)
 				if (MQTTProperties_hasProperty(&connack->properties, MQTTPROPERTY_CODE_SERVER_KEEP_ALIVE))
 				{
 					/* update the keep alive from the server keep alive */
-					int server_keep_alive = MQTTProperties_getNumericValue(&connack->properties, MQTTPROPERTY_CODE_SERVER_KEEP_ALIVE);
+					int server_keep_alive = (int)MQTTProperties_getNumericValue(&connack->properties, MQTTPROPERTY_CODE_SERVER_KEEP_ALIVE);
 					if (server_keep_alive != -999999)
 					{
 						Log(LOG_PROTOCOL, -1, "Setting keep alive interval to server keep alive %d", server_keep_alive);
@@ -2193,7 +2193,7 @@ thread_return_type WINAPI MQTTAsync_receiveThread(void* n)
 						{
 							if (MQTTProperties_hasProperty(&connack->properties, MQTTPROPERTY_CODE_RECEIVE_MAXIMUM))
 							{
-								int recv_max = MQTTProperties_getNumericValue(&connack->properties, MQTTPROPERTY_CODE_RECEIVE_MAXIMUM);
+								int recv_max = (int)MQTTProperties_getNumericValue(&connack->properties, MQTTPROPERTY_CODE_RECEIVE_MAXIMUM);
 								if (m->c->maxInflightMessages > recv_max)
 									m->c->maxInflightMessages = recv_max;
 							}
